fix(update-Records.py): create `add_repertory()`
Test for the existence of the `repertories` element on `record`; create
`repertories` if it does not exist.

Test whether a new repertory exists already as a child of `repertories`.

Call the function at appropriate points when adding MEC Bibliography
references and processing IMEV/NIMEV attribute values.

This fixes an error, by which the function `add_mec_refs()` inserted
`repertory` elements as direct children of `record` if called before
`extract_imev_etc()`.