fix(update-Records.py): repertories insertion
This fix is necessary because at least one record element has a comment
element at child-level, early in the sequence of children, and lxml
assigns index numbers to comment elements. Hence we cannot simply insert
the repertories element at index '2'.

Instead, we get the index of the alpha element and add 1.