refactor(scripts): rewrite transform-Bibl.py
Use lxml not xmltodict; improve structure and modularization.

The JSON output is essentially unchanged. The only difference is the
relative order of 'URL' and 'title' in four items.

The aim of refactoring was to check the logic, reduce script length, and
supply a better starting-point for transformation of PrintedBooks.xml
(which has a data structure similar to Bibliography.xml).

The script length is unchanged. I notice that I neglect to disaggregate
volume and number of journal publications. Fine. We'll fix that manually
in Zotero when we notice it.